deployment/requirements.txt
streamlit
scikit-learn
pandas
numpy
tensorflow-cpu